Audio, Mines, and stuff
Added:
Sustain Volume Slider (Careful with it, audio is not routed to it's own mixer channel yet, could vibrate the F out of your speakers. Also, 0 is mute, -1 to -5 is behaving as + value.)
Mine Detection (Error Checker)
Proper Audio Mixer (not all audio  routed to own channel yet)
Ability to swap hand color to colorless (keyboard key: 0).
Think that should be it.